+++ NuguAgents/Sources/CapabilityAgents/Session/SessionConst.swift
@@ -1,8 +1,8 @@
 //
-//  SessionManageable.swift
-//  NuguCore
+//  SessionConst.swift
+//  NuguAgents
 //
-//  Created by MinChul Lee on 2020/05/28.
+//  Created by MinChul Lee on 2020/06/17.
 //  Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,10 +20,6 @@
 
 import Foundation
 
-public protocol SessionManageable: class {
-    var syncedSessions: [Session] { get }
-    
-    func set(session: Session)
-    func sync(dialogRequestId: String)
-    func release(dialogRequestId: String)
+enum SessionConst {
+    static let sessionTimeout: DispatchTimeInterval = .seconds(60)
 }